fix(3ds-debug): use plain unsigned for the linear-heap snprintf args
u32 and uint32_t are unsigned long on devkitARM, so the %u conversions in
the DEBUG TOOLS memory line mismatched their arguments -- four
-Wformat= warnings. Same values at runtime, but only a DEBUG_TOOLS=1
build compiles that code path, which is why the default build was clean.

+++ platform/3ds/source/port_bottom_ui_3ds.c
@@ -3890,8 +3890,10 @@ static void RenderDebugToolsModal(int lang) {
          * anything else for an allocation failure. */
         PlatformGpu3DSStats st;
         PlatformGpu3DS_GetStats(&st);
-        const uint32_t freeKb = linearSpaceFree() / 1024u;
-        const uint32_t totalKb = st.linearHeapBytes / 1024u;
+        /* Plain unsigned, not u32/uint32_t: those are unsigned long on
+         * devkitARM and would not match snprintf's %u. */
+        const unsigned freeKb = (unsigned)(linearSpaceFree() / 1024u);
+        const unsigned totalKb = (unsigned)(st.linearHeapBytes / 1024u);
         char mem[48];
         snprintf(mem, sizeof(mem), "LINEAR %u.%01u / %u.%01u MB FREE",
                  freeKb / 1024u, (freeKb % 1024u) * 10u / 1024u,
